+++ src/main/java/com/tp_note/tp_note/service/ClientService.java
@@ -7,7 +7,7 @@ import com.tp_note.tp_note.ContratEntity;
 public interface ClientService {
 	
 	List getListeClients();
-	List getListeContrats(Long clientId);
-	List getContratDetails(Long cliendId, Long contratId);
-	List getClientsFromConseiller(Long conseillerId);
+	List getListeContrats(Integer clientId);
+	List getContratDetails(Integer cliendId, Integer contratId);
+	List getClientsFromConseiller(Integer conseillerId);
 }
